Set correct etype in NFS access rule (#11)
Currently an entity type `network` is used regardless of whether we're sharing a filesystem to an individual IP address or to a subnet.
This causes inconsistencies and errors in the Nexenta Fusion GUI - when we're inspecting and editing a share that has NFS access allowed, there are two fields for the IP and the netmask.

However the netmask field is empty when we shared to an individual IP, and making any changes to the share will fail, because the empty field won't pass validation [1]. We have to manually change the etype in GUI to "fqdn/ip".

This PR changes that behavior - we're setting `etype=network` if and only if we got a valid netmask from the client. Otherwise, we use `etype=fqdn`.

[1] https://pasteboard.co/Idrgag8.png