Place security context types into one dictionary instead of an array (#12)
Although I can't support this claim by any other reference, it looks like `securityContext` array entries in `PUT nas/nfs/{filesystem}` should in fact be dictionary keys. This is in spite of the fact that documentation says `securityContexts (array[nfsShare_securityContexts], optional): Security contexts,` - i.e. stating that `securityContexts` should be an array of contexts.

This has proven to be a problem empirically - when we define more than one access rule on a share via CLI, we suddenly can't see any rules in the Fusion GUI. When there's only one rule, we can see it correctly.

Inspecting the API responses for share access (`GET nas/nfs/{filesystem}`), we can see two different metadata structures depending on how we created those rules:
*Created via the Fusion GUI*, rules are split into buckets by their access level:
```
{
    "filesystem": "data/manila/share-52939911-b8d4-4581-96dc-92cb658f48ff",
    "shareState": "online",
    "mountPoint": "/data/manila/share-52939911-b8d4-4581-96dc-92cb658f48ff",
    "securityContexts": [
        {
            "readWriteList": [
                {
                    "allow": true,
                    "etype": "fqdn",
                    "entity": "10.1.31.217"
                },
                {
                    "allow": true,
                    "etype": "fqdn",
                    "entity": "10.1.31.218"
                }
            ],
           "readOnlyList": [
                {
                    "allow": true,
                    "etype": "fqdn",
                    "entity": "10.1.31.199"
                }
            ],
            "securityModes": [
                "sys"
            ]
        }
    ],
... omitted ...
}
```

*Created via `manila access-allow`*, the levels and the `securityModes` are duplicated:
```
{
    "filesystem": "data/manila/share-52939911-b8d4-4581-96dc-92cb658f48ff",
    "shareState": "online",
    "mountPoint": "/data/manila/share-52939911-b8d4-4581-96dc-92cb658f48ff",
    "securityContexts": [
        {
            "readWriteList": [
                {
                    "allow": true,
                    "etype": "fqdn",
                    "entity": "10.1.31.217"
                }
            ],
            "securityModes": [
                "sys"
            ]
        },
        {
            "readWriteList": [
                {
                    "allow": true,
                    "etype": "fqdn",
                    "entity": "10.1.31.218"
                }
            ],
            "securityModes": [
                "sys"
            ]
        },
        {
            "readOnlyList": [
                {
                    "allow": true,
                    "etype": "fqdn",
                    "entity": "10.1.31.199"
                }
            ],
            "securityModes": [
                "sys"
            ]
        }
    ],
... omitted ...
}
```

In the second case, managing the share via GUI is confusing because some users might have access to a share that otherwise appears not shared. Again, I have no reference which of these data structures is the correct one, I'm just assuming it's the first one based on the fact that Fusion expects it and that access rules keep working. Please try to verify this against some other reference if you have the possibility to, @Qeas .

With this in mind, the PR rearranges the metadata structure in the `PUT` call so that it's compliant with the Fusion API reader. I have verified that rules are set/unset the same way and the dataset is shared/unshared when we add the first rule/remove the last rule.